ram and swap fetchers impl

+++ src/crstat/mem/entities/swap_model.cr
@@ -1,11 +1,11 @@
 module CrStat::Mem::Entities
   struct SwapModel
-    property swap_total : Int32,
-      swap_free : Int32
+    property swap_total : Float64,
+      swap_free : Float64
 
     def initialize(
-      @swap_total : Int32,
-      @swap_free : Int32,
+      @swap_total : Float64,
+      @swap_free : Float64,
     )
     end
   end
